Rename RadicleSettings to RadicleGlobalSettings where we will have out application settings
